+++ src/common/game.d.ts
@@ -66,6 +66,7 @@ export interface GameState {
 
 export interface GameScore {
   cells: number,
+//   blanks: number,
   mines: number,
   level: number,
   mode?: PlayMode,